refactor(oracle): migrate tcp to per-protocol plugins
