+++ tools/oracle/tests/test_protocol_plugin_coverage.py
@@ -58,7 +58,6 @@ KNOWN_UNMIGRATED: frozenset[str] = frozenset(
         "rip",
         "ripng",
         "rsn",
-        "tcp",
     }
 )
